# [Kernel] support merge_attn_states CUDA kernel, 3x speedup