@DefTruth 2025-04-07
@WoosukKwon Hi~, Can you take a look at this PR? This PR use CUDA kernel instead of Triton to minimize CPU overhead. Compared to the Triton kernel, the CUDA kernel implemented in this PR can achieve a maximum speedup of over 3x.