@DefTruth 2025-04-08
@WoosukKwon Hi~ This PR is ready. Could you please take a look? Compared to the Triton kernel, the CUDA kernel implemented in this PR can achieve a maximum speedup of over `3x`.