@DefTruth 2025-04-08
@WoosukKwon Performance improved for R1 with PP=3 + TP=8 on L20,  4K IN:1K OUT (TTFT 5687.80 ms -> 5654.02 ms), 8K IN:64 OUT (TTFT  17673.93ms -> 14785.00ms), 16 concurrency.