@DefTruth 2025-04-10
[Lint and Deploy](https://github.com/vllm-project/vllm/actions/runs/14372017504/job/40296725721?pr=16173) CI failed, but seems not related to this PR. 

```bash
ERROR 04-10 04:06:36 [registry.py:345]     from torch._inductor.runtime.hints import DeviceProperties
ERROR 04-10 04:06:36 [registry.py:345]   File "/opt/venv/lib/python3.12/site-packages/torch/_inductor/runtime/hints.py", line 67, in <module>
ERROR 04-10 04:06:36 [registry.py:345]     from triton.compiler.compiler import AttrsDescriptor
ERROR 04-10 04:06:36 [registry.py:345] ImportError: cannot import name 'AttrsDescriptor' from 'triton.compiler.compiler' (/opt/venv/lib/python3.12/site-packages/triton/compiler/compiler.py)
```