@DefTruth 2025-04-11
@mgoin AMD build failed, should I add the #ifndef USE_ROCM macro restriction when binding in PyTorch?

```bash

#13 96.26 /app/vllm/csrc/torch_bindings.cpp:78:48: error: ‘merge_attn_states’ was not declared in this scope
--
  | #13 96.26    78 \|   ops.impl("merge_attn_states", torch::kCUDA, &merge_attn_states);
  | #13 96.26       \|                                                ^~~~~~~~~~~~~~~~~


```